First draft of the module based damage system

+++ BDArmory.Core/Events/DamageEventArgs.cs
@@ -8,7 +8,7 @@ namespace BDArmory.Core.Events
     {
         public int VesselId { get; set; }
         public int PartId { get; set; }
-        public double Damage { get; set; }
+        public float Damage { get; set; }
         public DamageOperation Operation { get; set; }
     }
 }
